payment and devcamp done

+++ StoreApp/StoreApp.Web/Models/Card.cs
@@ -23,7 +23,7 @@ public class Card
     {
         Items.RemoveAll(i => i.Product.Id == product.Id);
     }
-    public decimal CalculateTotal()
+    public double CalculateTotal()
     {
         return Items.Sum(i => i.Product.Price * i.Quantity);
     }
